Removed files and folders that should not be pushed to git, and added gitignore. Added a create user function. Changed the delete function in the file list page, to also delete the stored files and not only the database reference Changed file upload functionality to receive user and token in parameter instead of header. Changed Html templates and added some css for more understandable application layout

+++ photosec_demo/photos/views.py
@@ -2,10 +2,12 @@ import os
 from django.conf import settings
 from django.http import HttpResponse, HttpResponseNotFound, HttpResponseRedirect, JsonResponse
 from django.template import loader
-from django.contrib.auth import logout
-from django.contrib.auth.decorators import login_required
+from django.contrib.auth import logout, login
+from django.contrib.auth.decorators import login_required, user_passes_test
 from django.contrib.auth.models import User
+from django.contrib.auth.hashers import make_password
 from django.urls import reverse
+from django.core.exceptions import ObjectDoesNotExist
 from rest_framework.authtoken.models import Token
 from rest_framework.exceptions import ParseError
 from rest_framework.parsers import FileUploadParser
@@ -18,6 +20,34 @@ from .forms import FilesForm
 from .serializers import PhotoSerializer
 
 
+def create_user(request):
+    logout(request)
+
+    if request.method == 'GET':
+        template = loader.get_template('photos/create_user.html')
+        context = {
+        }
+        return HttpResponse(template.render(context, request))
+
+    if request.method == 'POST':
+        if User.objects.filter(username=request.POST.get('username')).exists():
+            template = loader.get_template('photos/create_user.html')
+            context = {
+               "error" : "username bestaat al"
+            }
+            return HttpResponse(template.render(context, request))
+        else:
+            user= User.objects.create(
+                username= request.POST.get('username'),
+                password= make_password(request.POST.get('password'))
+            )
+            user.save()
+            login(request, user)
+            return HttpResponseRedirect(reverse('file_list'))
+
+
+
+
 @login_required
 def qr_code(request):
     token, created = Token.objects.get_or_create(user=request.user)
@@ -27,7 +57,7 @@ def qr_code(request):
 
     template = loader.get_template('photos/qr_code.html')
     context = {
-        "user_id": request.user,
+        "user_id": request.user.id,
         "user_token": token.key,
         "app_id": "photosec_app",
     }
@@ -49,27 +79,36 @@ def file_list(request):
         files_list = Photos.objects.filter(user=request.user)
         for file in files_list:
             if request.POST.get(str(file.id))=='checked':
-                file.delete()
+                path = file.photo.path
+                if os.path.isfile(path):
+                    os.remove(path)
+                    file.delete()
         return HttpResponseRedirect(reverse('file_list'))
 
 
 class PhotoUploadView(APIView):
     parser_class = (FileUploadParser,)
 
-    authentication_classes = [TokenAuthentication]
-    permission_classes = [IsAuthenticated]
+    def post(self, request, user, token, format=None):
+        try:
+            token = Token.objects.get(pk=token)
+        except ObjectDoesNotExist:
+           return Response({'error': 'authentication failure'}, status=401)
 
-    def post(self, request, format=None):
-        if 'file' not in request.data:
-            raise ParseError('Empty content')
+        if token.user_id == user:
+            if 'file' not in request.data:
+                raise ParseError('Empty content')
 
-        file = request.data['file']
-        title = request.data['title']
+            file = request.data['file']
+            title = request.data['title']
 
-        photo = Photos(user=request.user, title=title, photo=file)
-        photo.save()
+            userId = User.objects.get(pk=user)
+            photo = Photos(user=userId, title=title, photo=file)
+            photo.save()
 
-        return HttpResponse('ok')
+            return HttpResponse('ok')
+        else:
+            return Response({'error': 'authentication failure'}, status=401)
 
 
 @login_required
@@ -81,11 +120,8 @@ def photos_retrieve(request):
 
         photos = Photos.objects.filter(user=request.user)
 
-        if not photos:
-            return HttpResponseNotFound('<h1>No photos found</h1>')
-        else:
-            serializer = PhotoSerializer(photos, many=True)
-            return JsonResponse(serializer.data, safe=False)
+        serializer = PhotoSerializer(photos, many=True)
+        return JsonResponse(serializer.data, safe=False)
     else:
         return HttpResponseNotFound('<h1>Not allowed</h1>')
